feat[L503C]:change boatGetTimes function define

+++ platform/Mobiletek-L503C-6S/src/osal/boatosalcommon.c
@@ -109,7 +109,7 @@ void BoatSleep(BUINT32 second)
 * @return
 ****************************************************************************************
 */
-long int BoatGetTimes()
+uint32_t BoatGetTimes()
 {
     uint32_t  time;
     uint32_t ret = ol_time(&time);
